# 🚀 Use native URL() on module builds