--- a/src/url.js
+++ b/src/url.js
@@ -92,9 +92,14 @@ export function getWinOrigin(win) {
  * testing by freezing the object.
  * @param {string} url
  * @param {boolean=} opt_nocache
+ *   Cache is altogether ignored on ESM builds, since an <a> is not used.
  * @return {!Location}
  */
 export function parseUrlDeprecated(url, opt_nocache) {
+  if (IS_ESM) {
+    return /** @type {!Location} */ (new URL(url));
+  }
+
   if (!a) {
     a = /** @type {!HTMLAnchorElement} */ (self.document.createElement('a'));
     cache = self.__AMP_URL_CACHE || (self.__AMP_URL_CACHE = new LruCache(100));
@@ -111,10 +116,15 @@ export function parseUrlDeprecated(url, opt_nocache) {
  * @param {!HTMLAnchorElement} a
  * @param {string} url
  * @param {LruCache=} opt_cache
+ *   Cache is altogether ignored on ESM builds, since an <a> is not used.
  * @return {!Location}
  * @restricted
  */
 export function parseUrlWithA(a, url, opt_cache) {
+  if (IS_ESM) {
+    return /** @type {!Location} */ (new URL(url));
+  }
+
   if (opt_cache && opt_cache.has(url)) {
     return opt_cache.get(url);
   }